Summary: Update config and add server load to config
Description: This change updated the configuration to remove unnecessary log statements, removed the use of `private-keys` from the config, updated the `server-load` to be a configurable option in the config, and added a check for the existence of the `global_config.yml` file in the `config.NewConfig()` function. Additionally, this change added the ability to update the config in the `main.go` file to support the new `server-load` option. Finally, this change replaced the use of `node.NewNode()` with `NewNodeView()` in the `UpdateNode()` function in the `bulletin_board.go` file to allow for better handling of node updates.