Improve client-side code and add PrettyLogger to api/logger.go, updating metrics.go and clientStatus.go.
This change fixed a bug by adding PrettyLogger to api/logger.go, which in turn improved code in metrics.go and clientStatus.go. The client code in client.go was also updated to generate dummy messages and add slog.Info logging. The index.html file was updated to sort messages based on the concatenation of From and To strings, as well as the message itself.