Update visualization page to draw arrows between nodes and highlight cell on hover.
This change updated the visualization pages index.html file to draw arrows between nodes and highlight the cells on hover. The code now includes functions to draw arrows using CanvasRenderingContext2D and set the background color of cells to yellow when hovered over. It also gets the bounding rectangles of cells to accurately place the arrow. However, the drawing of the arrow was commented out, suggesting that it is intended to be implemented in the future.