Update metrics package and plugin addresses to use structs API.
This change updates the metrics package and plugin addresses to use structs API, replacing the previous api.NodeStatus and api.ClientStatus with structs.NodeStatus and structs.ClientStatus respectively. The file config/config.go is also modified to set the addresses of nodes, clients, bulletin board, and metrics. Additionally, the function AddressToName is introduced to map addresses to names.