Update client and API function timeouts and add retries for sending onions.
This change updates the timeout for sending onions in the client and API functions. The timeout for sending onions has been increased from 10 seconds to 30 seconds. Additionally, the code that processes and sends onions has been refactored to use goroutines for asynchronous sending and re-try in case of errors. This should improve reliability and performance of the onion sending process.